Finish exercise two
I used a grouping selector to apply the 24px font-size
declaration to the third and fourth element instead of adding a
class to the fourth element that it shares with the third element.
My way moves more complexity to the CSS file, but doesn't require
adding another class. Don't know if it is better or worse.