Finish exercise 6
Preferred moving elements around, used a descendant combinator
for the last fix. Alternative which I didn't want to implement
was to just fluff the specificity by duplicating the .child tag,
so '.child.child' which would give it a specificity of 0-2-0, beating
the 0-1-1 of the 'div.text' selector. Felt like a hack, so chose not to.