Finish flex exercise 4
Even more flexibility in how to approach. The solution used a clever
thing: It applied 'text-align: center' to the body element, so that it'd be
inherited by everything that needs to be centered (h1, the image, text).
I applied those separately and made the image a block-level element so
that I could center it with margin.